[feat] #17 캐릭터 선택 다이얼로그 구현
