+++ src/styles/css/CharacterDialog.css
@@ -8,22 +8,110 @@
   background: white;
   padding: 1.5rem;
   border-radius: 0.5rem;
-  width: 300px;
+  width: 340px;
+  height: 562px;
   max-width: 80%;
+  overflow-y: auto;
   position: fixed;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
 }
-
-.dialog-title {
+.dialog-content .dialog-title {
   font-size: 1.25rem;
   font-weight: bold;
   margin-bottom: 0.5rem;
 }
-
-.dialog-close {
+.dialog-content .dialog-close {
   background: #ccc;
   padding: 0.25rem 0.75rem;
   border-radius: 0.25rem;
-}/*# sourceMappingURL=CharacterDialog.css.map */
+}
+.dialog-content .scroll-wrapper {
+  position: relative;
+  height: 276px;
+  overflow-y: auto;
+  margin: 16px 0;
+  padding: 0px;
+  z-index: 1;
+  background: linear-gradient(to top, white, rgba(255, 255, 255, 0));
+}
+.dialog-content .scroll-gradient {
+  position: sticky;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  height: 20px;
+  z-index: 2;
+  pointer-events: none;
+  background: linear-gradient(to top, white, rgba(255, 255, 255, 0));
+}
+.dialog-content .character-grid {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 20px 6px;
+  margin: 16px 0;
+}
+.dialog-content .character-grid .character-card {
+  border: 1px solid #eee;
+  border-radius: 8px;
+  padding: 4px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  cursor: pointer;
+  transition: 0.2s;
+}
+.dialog-content .character-grid .character-card .character-name {
+  font-size: 0.875rem;
+  margin-top: 5px;
+}
+.dialog-content .character-grid .character-card img {
+  width: 44px;
+  height: 44px;
+}
+.dialog-content .dialog-bottom-fieldset {
+  all: unset;
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  align-items: center;
+  margin-top: 50px;
+}
+.dialog-content .dialog-bottom-fieldset .dialog-bottom-label {
+  font-size: 12px;
+  line-height: 18px;
+  font-weight: 500;
+  letter-spacing: -2.5%;
+  color: #767676;
+}
+.dialog-content .dialog-bottom-fieldset .dialog-bottom-input {
+  all: unset;
+  width: 100%;
+  height: 46px;
+  padding: 0 16px;
+  border-radius: 8px;
+  font-size: 1rem;
+  font-weight: 500;
+  color: #333;
+  border: 1px solid #ccc;
+  box-sizing: border-box;
+}
+.dialog-content .dialog-bottom-fieldset .dialog-bottom-input:focus {
+  outline: none;
+  border-color: #111111;
+}
+.dialog-content .dialog-close {
+  all: unset;
+  width: 100%;
+  height: 46px;
+  background-color: #FFE788;
+  color: #111111;
+  border-radius: 8px;
+  margin-top: 28px;
+  font-size: 1rem;
+  font-weight: 500;
+  text-align: center;
+  line-height: 46px;
+  cursor: pointer;
+}
